Executing Applkcaiton with compt
