+++ src/MigrationTools.ConsoleUI/MigrationTools.ConsoleUI.csproj
@@ -6,7 +6,7 @@
     <Version>0.0.0.0</Version>
     <Authors>Martin Hinshelwood</Authors>
     <Company>naked Agility with Martin Hinshelwood</Company>
-    <Product>Azure DevOps Migration Tools</Product>
+    <Product>Azure DevOps Migration Tools [REST API]</Product>
     <AssemblyName>migration</AssemblyName>
   </PropertyGroup>
