+++ src/MigrationTools.Sinks.TfsObjectModel/packages.config
@@ -15,6 +15,7 @@
   <package id="Microsoft.VisualStudio.Services.Client" version="16.153.0" targetFramework="net472" />
   <package id="Microsoft.VisualStudio.Services.InteractiveClient" version="16.153.0" targetFramework="net472" />
   <package id="Newtonsoft.Json" version="10.0.3" targetFramework="net472" />
+  <package id="Serilog" version="2.10.0" targetFramework="net472" />
   <package id="System.Buffers" version="4.5.1" targetFramework="net472" />
   <package id="System.Collections.Immutable" version="1.4.0" targetFramework="net472" />
   <package id="System.Diagnostics.DiagnosticSource" version="4.6.0" targetFramework="net472" />
