+++ src/MigrationTools/CustomDiagnostics/CustomConverter.cs
