+++ src/VstsSyncMigrator.Console/appsettings.json
@@ -1,0 +1,5 @@
+{
+  "ApplicationInsights": {
+    "InstrumentationKey": "4b9bb17b-c7ee-43e5-b220-ec6db2c33373"
+  }
+}
