+++ src/VstsSyncMigrator.Console/packages.config
@@ -26,6 +26,7 @@
   <package id="Microsoft.Diagnostics.Tracing.EventRegister" version="1.1.28" targetFramework="net45" />
   <package id="Microsoft.Diagnostics.Tracing.EventSource" version="1.1.28" targetFramework="net45" />
   <package id="Microsoft.Diagnostics.Tracing.EventSource.Redist" version="2.2.0" targetFramework="net461" />
+  <package id="Microsoft.DotNet.PlatformAbstractions" version="2.0.4" targetFramework="net461" />
   <package id="Microsoft.Extensions.Configuration" version="3.1.8" targetFramework="net461" />
   <package id="Microsoft.Extensions.Configuration.Abstractions" version="3.1.8" targetFramework="net461" />
   <package id="Microsoft.Extensions.Configuration.Binder" version="3.1.8" targetFramework="net461" />
@@ -36,6 +37,7 @@
   <package id="Microsoft.Extensions.Configuration.UserSecrets" version="3.1.8" targetFramework="net461" />
   <package id="Microsoft.Extensions.DependencyInjection" version="3.1.8" targetFramework="net461" />
   <package id="Microsoft.Extensions.DependencyInjection.Abstractions" version="3.1.8" targetFramework="net461" />
+  <package id="Microsoft.Extensions.DependencyModel" version="2.0.4" targetFramework="net461" />
   <package id="Microsoft.Extensions.FileProviders.Abstractions" version="3.1.8" targetFramework="net461" />
   <package id="Microsoft.Extensions.FileProviders.Physical" version="3.1.8" targetFramework="net461" />
   <package id="Microsoft.Extensions.FileSystemGlobbing" version="3.1.8" targetFramework="net461" />
@@ -69,8 +71,11 @@
   <package id="NuGet.Core" version="2.14.0" targetFramework="net45" />
   <package id="NuGet.Versioning" version="5.7.0" targetFramework="net461" />
   <package id="Serilog" version="2.10.0" targetFramework="net461" />
+  <package id="Serilog.Enrichers.Environment" version="2.1.3" targetFramework="net461" />
+  <package id="Serilog.Enrichers.Process" version="2.0.1" targetFramework="net461" />
   <package id="Serilog.Extensions.Hosting" version="3.1.0" targetFramework="net461" />
   <package id="Serilog.Extensions.Logging" version="3.0.1" targetFramework="net461" />
+  <package id="Serilog.Settings.Configuration" version="3.1.0" targetFramework="net461" />
   <package id="Serilog.Sinks.ApplicationInsights" version="3.1.0" targetFramework="net461" />
   <package id="Serilog.Sinks.Console" version="3.1.1" targetFramework="net461" />
   <package id="Serilog.Sinks.File" version="4.1.0" targetFramework="net461" />
